Move SpatialJoin prefilter test helpers to own file
Preparation for tests of #2246
